requirements.txt
numpy
pandas
wandb
matplotlib
tqdm
scikit-learn
facenet-pytorch
Pillow
torch
torchvision
torchaudio
# Use the extra index URL for PyTorch CUDA wheels
--extra-index-url https://download.pytorch.org/whl/cu128
